👌Separação e criação de Summary para diferentes tipos de empréstimos

+++ spring-basics/src/main/java/com/example/spring_basics/dto/response/loan/LoanResponseDTO.java
@@ -12,7 +12,7 @@ public record LoanResponseDTO(
         LocalDate loanDate,
         LocalDate dueDate,
         LocalDate returnDate,
-        Integer fineAmount,
+        double fineAmount,
         LoanStatus status,
         BookCopySummaryDTO bookCopy,
         UserSummaryDTO user) {
